apps/api/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
click==8.3.1
colorama==0.4.6
fastapi==0.121.3
-e ../..
h11==0.16.0
httptools==0.7.1
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
pydantic==2.12.4
pydantic-settings==2.12.0
pydantic_core==2.41.5
pymongo==4.7.3
Pygments==2.19.2
python-dotenv==1.2.1
python-multipart==0.0.20
PyYAML==6.0.3
rich==14.2.0
scipy==1.16.3
shellingham==1.5.4
sniffio==1.3.1
starlette==0.50.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.38.0
watchfiles==1.1.1
websockets==15.0.1
